Added stack allocator class

+++ BBEngine/BBEngine/Allocators/Allocator.h
@@ -13,7 +13,7 @@ namespace BBE {
 			~Allocator() = default;
 
 			virtual void Init(const size_t& size) noexcept = 0;
-			virtual void* Alloc(const uint32_t& a_Size, const uint32_t& a_Align = DEFAULT_ALIGNMENT) = 0;
+			virtual void* Alloc(const size_t& a_Size, const size_t& a_Align = DEFAULT_ALIGNMENT) = 0;
 			virtual void* Realloc(void* a_OldData, const size_t& a_OldSize, const size_t& a_NewSize, const size_t& a_Align = DEFAULT_ALIGNMENT) = 0;
 			virtual void Free(void* ptr) = 0;
 			virtual void Clear() = 0;
